Routine plan form depending on create/update
